:rocket: [Feat] 글 작성 시 유저 아이디 읽기만 가능 및 공지사항 관리자 권한일 경우에만 보기 기능 구현

+++ trip-sketch/src/components/board/BoardDelete.vue
@@ -1,42 +1,54 @@
 <template>
-	<div class="regist">
-		<h1>SSAFY 글 삭제</h1>
-		<div>삭제중....</div>
-	</div>
+  <div class="regist">
+    <h1>SSAFY 글 삭제</h1>
+    <div>삭제중....</div>
+  </div>
 </template>
 
 <script>
 import axios from "axios";
+import { mapState } from "vuex";
+const memberStore = "memberStore";
 const url = "http://localhost:80/board/notice/";
 export default {
-	name: "BoardDelete",
-	data() {
-		return {
-			articleId: Number,
-		};
-	},
-	async created() {
-		// 비동기
-		// TODO : 글번호에 해당하는 글을 삭제.
-		this.articleId = this.$route.params.articleId;
-		// console.log(this.articleId + "번글 삭제!!!!!");
-		// this.articles = JSON.parse(localStorage.getItem("articles"));
-		// this.articles = this.articles.filter((article) => article.articleId !== this.articleId);
-		// localStorage.setItem("articles", JSON.stringify(this.articles));
-		await axios.delete(url + this.articleId + "/delete");
-		this.moveList();
-	},
-	methods: {
-		moveList() {
-			//   console.log("글목록 보러가자!!!");
-			this.$router.push({ path: "list" });
-		},
-	},
+  name: "BoardDelete",
+  data() {
+    return {
+      articleId: Number,
+      authorization: "",
+    };
+  },
+  computed: {
+    ...mapState(memberStore, ["isLogin", "isLoginError", "userInfo"]),
+  },
+  async created() {
+    this.authorization = this.userInfo.authorization;
+    if (this.authorization != "a") {
+      alert("글 삭제 권한이 없습니다!");
+      this.$router.push({ path: "list" });
+      return;
+    }
+    // 비동기
+    // TODO : 글번호에 해당하는 글을 삭제.
+    this.articleId = this.$route.params.articleId;
+    // console.log(this.articleId + "번글 삭제!!!!!");
+    // this.articles = JSON.parse(localStorage.getItem("articles"));
+    // this.articles = this.articles.filter((article) => article.articleId !== this.articleId);
+    // localStorage.setItem("articles", JSON.stringify(this.articles));
+    await axios.delete(url + this.articleId + "/delete");
+    this.moveList();
+  },
+  methods: {
+    moveList() {
+      //   console.log("글목록 보러가자!!!");
+      this.$router.push({ path: "list" });
+    },
+  },
 };
 </script>
 
 <style scoped>
 * {
-	color: black;
+  color: black;
 }
 </style>
